requirements.txt

Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn==22.0.0
psutil==6.0.0
APScheduler==3.10.4
qrcode==7.4.2
pillow==10.4.0
requests==2.32.3
ipaddress==1.0.23
